enable Paging Mode feature for all architectures
