+++ common/protos/limine.c
@@ -438,13 +438,12 @@ noreturn void limine_load(char *config, char *cmdline) {
 #define paging_mode_vmm_to_limine(x) (x)
 #endif
 
-#if defined(__riscv64)
+    bool have_paging_mode_request = false;
 FEAT_START
-
-
     struct limine_paging_mode_request *pm_request = get_request(LIMINE_PAGING_MODE_REQUEST);
     if (pm_request == NULL)
         break;
+    have_paging_mode_request = true;
 
     if (pm_request->mode > LIMINE_PAGING_MODE_MAX) {
         print("warning: ignoring invalid mode in paging mode request\n");
@@ -463,14 +462,16 @@ FEAT_START
 
 FEAT_END
 
-#else
-
     // 5 level paging feature & HHDM slide
 FEAT_START
     struct limine_5_level_paging_request *lv5pg_request = get_request(LIMINE_5_LEVEL_PAGING_REQUEST);
     if (lv5pg_request == NULL)
         break;
 
+    if (have_paging_mode_request) {
+        print("paging: ignoring 5-level paging request in favor of paging mode request\n");
+        break;
+    }
 #if defined (__x86_64__) || defined (__i386__)
     if (max_paging_mode < PAGING_MODE_X86_64_5LVL)
         break;
@@ -479,6 +480,8 @@ FEAT_START
     if (max_paging_mode < PAGING_MODE_AARCH64_5LVL)
         break;
     paging_mode = PAGING_MODE_AARCH64_5LVL;
+#elif defined (__riscv64)
+    print("warning: the 5-level paging request is not supported on RISC-V\n");
 #else
 #error Unknown architecture
 #endif
@@ -486,8 +489,6 @@ FEAT_START
     void *lv5pg_response = ext_mem_alloc(sizeof(struct limine_5_level_paging_response));
     lv5pg_request->response = reported_addr(lv5pg_response);
 FEAT_END
-
-#endif
 
     direct_map_offset = paging_mode_higher_half(paging_mode);
     if (kaslr) {
